refactor: split variables into package
